complete refactor with spectator
