+++ packages/webapp/src/inductions/components/induction-lists/invitee-inductions.tsx
@@ -4,7 +4,7 @@ import * as InductionTable from "_app/ui/table";
 import { getInductionRemainingTimeDays } from "../../utils";
 import { Induction, InductionRole } from "../../interfaces";
 import { InductionStatusButton } from "./induction-status-button";
-import { EndorsersNames } from "./endorsers-names";
+import { AccountName, EndorsersNames } from "./induction-names";
 
 interface Props {
     inductions: Induction[];
@@ -41,26 +41,21 @@ const INVITEE_INDUCTION_COLUMNS: InductionTable.Column[] = [
 ];
 
 const getTableData = (inductions: Induction[]): InductionTable.Row[] => {
-    return inductions.map((induction) => {
-        const { data: inviter } = useMemberByAccountName(induction.inviter);
-        const remainingTime = getInductionRemainingTimeDays(induction);
-
-        return {
-            key: induction.id,
-            inviter: inviter ? inviter.name : induction.inviter,
-            witnesses: (
-                <EndorsersNames
-                    induction={induction}
-                    skipEndorser={induction.inviter}
-                />
-            ),
-            time_remaining: remainingTime,
-            status: (
-                <InductionStatusButton
-                    induction={induction}
-                    role={InductionRole.Invitee}
-                />
-            ),
-        };
-    });
+    return inductions.map((induction) => ({
+        key: induction.id,
+        inviter: <AccountName account={induction.inviter} />,
+        witnesses: (
+            <EndorsersNames
+                induction={induction}
+                skipEndorser={induction.inviter}
+            />
+        ),
+        time_remaining: getInductionRemainingTimeDays(induction),
+        status: (
+            <InductionStatusButton
+                induction={induction}
+                role={InductionRole.Invitee}
+            />
+        ),
+    }));
 };
